pyproject.toml
[project]
name = "mutual-fund-brokerage-api"
version = "0.1.0"
description = ""
authors = [
    {name = "sujeet-agrahari",email = "sujeet.rgcsm@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.10,<0.116.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
]

[tool.poetry]
packages = [{include = "mutual_fund_brokerage_api", from = "src"}]

[tool.poetry.scripts]
start = "mutual_fund_brokerage_api.server:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
